Move the base handler out into its own module.
